🐛 Fix.  Reorganising the user page widget
