+++ lib/pages/9.pages.dart
@@ -3,4 +3,6 @@ export 'package:mobile1_flutter_coding_test/pages/main_page.dart';
 export 'package:mobile1_flutter_coding_test/pages/room_list_page.dart';
 export 'package:mobile1_flutter_coding_test/pages/user/user_list_page.dart';
 export 'package:mobile1_flutter_coding_test/pages/user/user_card_widget.dart';
+export 'package:mobile1_flutter_coding_test/pages/user/user_info_widget.dart';
+export 'package:mobile1_flutter_coding_test/pages/user/user_photo_widget.dart';
 export 'package:mobile1_flutter_coding_test/pages/error/error_page.dart';
